% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific-package.R
\docType{package}
\name{scientific-package}
\alias{scientific}
\alias{scientific-package}
\title{scientific: Scientific Themed Reporting for R Markdown}
\description{
Offers R Markdown output formats designed with various scientific styles, allowing users to generate PDF and HTML outputs. The output has a contemporary appearance with vibrant visuals, providing numerous styles for effective highlighting. The package also includes additional features specifically tailored for front-page slides, enhancing the overall presentation and customization options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scientific.obi.obianom.com}
  \item Report bugs at \url{https://github.com/oobianom/scientific/issues}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
