% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific-package.R
\docType{package}
\name{scientific-package}
\alias{scientific}
\alias{scientific-package}
\title{scientific: Scientific themed reporting for R Markdown}
\description{
Provides R Markdown output formats to use scientific styles for PDF and HTML output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scientific.obi.obianom.com}
  \item Report bugs at \url{https://github.com/oobianom/scientific/issues}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
