% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific-package.R
\docType{package}
\name{scientific-package}
\alias{scientific}
\alias{scientific-package}
\title{scientific: Two Highly Customizable 'rmarkdown' Themes for Scientific Reports}
\description{
Offers 'markdown' output formats designed with various styles, allowing users to generate HTML reports tailored for scientific or machine learning showcase. The output has a contemporary appearance with vibrant visuals, providing numerous styles for effective highlighting. Created using the 'tufte' \url{https://rstudio.github.io/tufte/} package code as a starting point.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scientific.obi.obianom.com}
  \item Report bugs at \url{https://github.com/oobianom/scientific/issues}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
